printing the size of variable types

+++ 0x00-hello_world/6-size.c
@@ -10,11 +10,11 @@ int main(void)
 	long int l;
 	long long int L;
 	float f;
-printf("Size of char: %d byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of int: %d byte(s)\n", (unsigned long)sizeof(int));
-	printf("Size of long int: %d byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of long long int: %d byte(s)\n", (unsigned long)sizeof(long long int);
-	printf("Size of float: %d byte(s)\n", (unsigned long)sizeof(float));
+printf("Size of char: %d byte(s)\n", sizeof(char));
+	printf("Size of int: %d byte(s)\n", sizeof(int));
+	printf("Size of long int: %d byte(s)\n", sizeof(long int));
+	printf("Size of long long int: %d byte(s)\n", sizeof(long long int);
+	printf("Size of float: %d byte(s)\n", sizeof(float));
 	return (0);
 }
